--- a/curves/src/pasta/curves/tests.rs
+++ b/curves/src/pasta/curves/tests.rs
@@ -1,5 +1,5 @@
 use ark_algebra_test_templates::{curves::*, groups::*};
-use ark_ec::AffineCurve;
+use ark_ec::AffineRepr;
 use ark_std::test_rng;
 use rand::Rng;