--- a/curves/src/pasta/fields/fq.rs
+++ b/curves/src/pasta/fields/fq.rs
@@ -1,46 +1,53 @@
-use ark_ff::{
-    biginteger::BigInteger256 as BigInteger, FftParameters, Fp256, Fp256Parameters, FpParameters,
-};
+use super::fft::{FftParameters, Fp256Parameters, FpParameters};
+use ark_ff::{biginteger::BigInteger256 as BigInteger, Fp256};
 
 pub struct FqParameters;
 
-pub type Fq = Fp256<FqParameters>;
+use ark_ff::fields::{MontBackend, MontConfig};
+
+#[derive(MontConfig)]
+#[modulus = "28948022309329048855892746252171976963363056481941647379679742748393362948097"]
+#[generator = "5"]
+pub struct FrConfig;
+pub type Fq = Fp256<MontBackend<FrConfig, 4>>;
 
 impl Fp256Parameters for FqParameters {}
+
 impl FftParameters for FqParameters {
     type BigInt = BigInteger;
 
     const TWO_ADICITY: u32 = 32;
 
     #[rustfmt::skip]
-    const TWO_ADIC_ROOT_OF_UNITY: BigInteger = BigInteger([
+    const TWO_ADIC_ROOT_OF_UNITY: BigInteger = BigInteger::new([
         0x218077428c9942de, 0xcc49578921b60494, 0xac2e5d27b2efbee2, 0xb79fa897f2db056
     ]);
 }
+
 impl FpParameters for FqParameters {
     // 28948022309329048855892746252171976963363056481941647379679742748393362948097
-    const MODULUS: BigInteger = BigInteger([
+    const MODULUS: BigInteger = BigInteger::new([
         0x8c46eb2100000001,
         0x224698fc0994a8dd,
         0x0,
         0x4000000000000000,
     ]);
 
-    const R: BigInteger = BigInteger([
+    const R: BigInteger = BigInteger::new([
         0x5b2b3e9cfffffffd,
         0x992c350be3420567,
         0xffffffffffffffff,
         0x3fffffffffffffff,
     ]);
 
-    const R2: BigInteger = BigInteger([
+    const R2: BigInteger = BigInteger::new([
         0xfc9678ff0000000f,
         0x67bb433d891a16e3,
         0x7fae231004ccf590,
         0x96d41af7ccfdaa9,
     ]);
 
-    const MODULUS_MINUS_ONE_DIV_TWO: BigInteger = BigInteger([
+    const MODULUS_MINUS_ONE_DIV_TWO: BigInteger = BigInteger::new([
         0xc623759080000000,
         0x11234c7e04ca546e,
         0x0,
@@ -49,13 +56,13 @@ impl FpParameters for FqParameters {
 
     // T and T_MINUS_ONE_DIV_TWO, where MODULUS - 1 = 2^S * T
 
-    const T: BigInteger = BigInteger([0x994a8dd8c46eb21, 0x224698fc, 0x0, 0x40000000]);
+    const T: BigInteger = BigInteger::new([0x994a8dd8c46eb21, 0x224698fc, 0x0, 0x40000000]);
 
     const T_MINUS_ONE_DIV_TWO: BigInteger =
-        BigInteger([0x4ca546ec6237590, 0x11234c7e, 0x0, 0x20000000]);
+        BigInteger::new([0x4ca546ec6237590, 0x11234c7e, 0x0, 0x20000000]);
 
     // GENERATOR = 5
-    const GENERATOR: BigInteger = BigInteger([
+    const GENERATOR: BigInteger = BigInteger::new([
         0x96bc8c8cffffffed,
         0x74c2a54b49f7778e,
         0xfffffffffffffffd,