--- a/curves/src/pasta/fields/mod.rs
+++ b/curves/src/pasta/fields/mod.rs
@@ -4,5 +4,7 @@ pub use self::fp::*;
 pub mod fq;
 pub use self::fq::*;
 
+pub mod fft;
+
 #[cfg(test)]
 mod tests;