--- a/curves/src/pasta/fields/mod.rs
+++ b/curves/src/pasta/fields/mod.rs
@@ -1,3 +1,4 @@
+use ark_ff::Field;
 pub mod fp;
 pub use self::fp::*;
 
@@ -5,6 +6,43 @@ pub mod fq;
 pub use self::fq::*;
 
 pub mod fft;
+
+#[derive(Debug, PartialEq)]
+pub enum LegendreSymbol {
+    Zero = 0,
+    QuadraticResidue = 1,
+    QuadraticNonResidue = -1,
+}
+
+impl LegendreSymbol {
+    pub fn is_zero(&self) -> bool {
+        *self == LegendreSymbol::Zero
+    }
+
+    pub fn is_qnr(&self) -> bool {
+        *self == LegendreSymbol::QuadraticNonResidue
+    }
+
+    pub fn is_qr(&self) -> bool {
+        *self == LegendreSymbol::QuadraticResidue
+    }
+}
+
+/// The interface for a field that supports an efficient square-root operation.
+pub trait SquareRootField: Field {
+    /// Returns a `LegendreSymbol`, which indicates whether this field element is
+    ///  1 : a quadratic residue
+    ///  0 : equal to 0
+    /// -1 : a quadratic non-residue
+    fn legendre(&self) -> LegendreSymbol;
+
+    /// Returns the square root of self, if it exists.
+    #[must_use]
+    fn sqrt(&self) -> Option<Self>;
+
+    /// Sets `self` to be the square root of `self`, if it exists.
+    fn sqrt_in_place(&mut self) -> Option<&mut Self>;
+}
 
 #[cfg(test)]
 mod tests;