--- a/kimchi/src/circuits/domain_constant_evaluation.rs
+++ b/kimchi/src/circuits/domain_constant_evaluation.rs
@@ -2,8 +2,8 @@
 
 use crate::circuits::domains::EvaluationDomains;
 use ark_ff::FftField;
+use ark_poly::DenseUVPolynomial;
 use ark_poly::EvaluationDomain;
-use ark_poly::UVPolynomial;
 use ark_poly::{univariate::DensePolynomial as DP, Evaluations as E, Radix2EvaluationDomain as D};
 use serde::{Deserialize, Serialize};
 use serde_with::serde_as;