… Cargo.lock
… book/src/specs/kimchi.md
… circuit-construction/Cargo.toml
… circuit-construction/src/tests/example_proof.rs
… circuit-construction/src/writer.rs
… curves/Cargo.toml
… groupmap/Cargo.toml
… groupmap/src/lib.rs
… groupmap/tests/groupmap.rs
… hasher/Cargo.toml
… hasher/src/roinput.rs
… kimchi/Cargo.toml
… kimchi/src/circuits/expr.rs
… kimchi/src/circuits/gate.rs
… kimchi/src/circuits/polynomials/and.rs
… kimchi/src/circuits/polynomials/endomul_scalar.rs
… kimchi/src/circuits/polynomials/foreign_field_add/gadget.rs
… kimchi/src/circuits/polynomials/foreign_field_mul/gadget.rs
… kimchi/src/circuits/polynomials/keccak/gadget.rs
… kimchi/src/circuits/polynomials/permutation.rs
… kimchi/src/circuits/polynomials/poseidon.rs
… kimchi/src/circuits/polynomials/range_check/gadget.rs
… kimchi/src/circuits/polynomials/rot.rs
… kimchi/src/circuits/polynomials/turshi.rs
… kimchi/src/circuits/polynomials/xor.rs
… kimchi/src/circuits/wires.rs
… kimchi/src/circuits/witness/mod.rs
… kimchi/src/curve.rs
… kimchi/src/folding/error_term.rs
… kimchi/src/folding/expressions.rs