Benchmark:
```shell
cargo criterion -p kimchi --bench proof_criterion
```

Before:
```
Proof creation/proof creation (SRS size 2^{16}, 1014 gates)                                                                          
                        time:   [852.95 ms 859.61 ms 867.67 ms]
                        change: [+0.3236% +1.3483% +2.5489%] (p = 0.03 < 0.05)
                        Change within noise threshold.
Benchmarking Proof creation/proof creation (SRS size 2^{16}, 16374 gates): Warming up for 3.0000 s
Warning: Unable to complete 10 samples in 5.0s. You may wish to increase target time to 16.9s.
Proof creation/proof creation (SRS size 2^{16}, 16374 gates)                                                                          
                        time:   [1.6794 s 1.6851 s 1.6919 s]
                        change: [-5.7030% -5.2167% -4.7135%] (p = 0.00 < 0.05)
                        Performance has improved.

Benchmarking Proof verification/proof verification (SRS size 2^{16}, 1014 gates): Warming up for 3.0000 s
Warning: Unable to complete 100 samples in 5.0s. You may wish to increase target time to 6.9s, or reduce sample count to 70.
Proof verification/proof verification (SRS size 2^{16}, 1014 gates)                                                                            
                        time:   [67.749 ms 68.204 ms 68.678 ms]
                        change: [-7.2764% -6.2600% -5.2345%] (p = 0.00 < 0.05)
                        Performance has improved.
Benchmarking Proof verification/proof verification (SRS size 2^{16}, 16374 gates): Warming up for 3.0000 s
Warning: Unable to complete 100 samples in 5.0s. You may wish to increase target time to 6.9s, or reduce sample count to 70.
Proof verification/proof verification (SRS size 2^{16}, 16374 gates)                                                                            
                        time:   [67.834 ms 68.301 ms 68.802 ms]
                        change: [-7.5480% -6.5346% -5.4975%] (p = 0.00 < 0.05)
                        Performance has improved.
```
After (commit 70a1d2325dd2ec9dffeba96acfed6401f4ea42f5):
```
Benchmarking Proof creation/proof creation (SRS size 2^{16}, 1014 gates): Warming up for 3.0000 s
Warning: Unable to complete 10 samples in 5.0s. You may wish to increase target time to 8.4s.
Proof creation/proof creation (SRS size 2^{16}, 1014 gates)                                                                          
                        time:   [842.78 ms 848.18 ms 854.01 ms]
Benchmarking Proof creation/proof creation (SRS size 2^{16}, 16374 gates): Warming up for 3.0000 s
Warning: Unable to complete 10 samples in 5.0s. You may wish to increase target time to 17.7s.
Proof creation/proof creation (SRS size 2^{16}, 16374 gates)                                                                          
                        time:   [1.7711 s 1.7778 s 1.7846 s]

Benchmarking Proof verification/proof verification (SRS size 2^{16}, 1014 gates): Warming up for 3.0000 s
Warning: Unable to complete 100 samples in 5.0s. You may wish to increase target time to 7.3s, or reduce sample count to 60.
Proof verification/proof verification (SRS size 2^{16}, 1014 gates)                                                                            
                        time:   [72.151 ms 72.759 ms 73.402 ms]
Benchmarking Proof verification/proof verification (SRS size 2^{16}, 16374 gates): Warming up for 3.0000 s
Warning: Unable to complete 100 samples in 5.0s. You may wish to increase target time to 7.3s, or reduce sample count to 60.
Proof verification/proof verification (SRS size 2^{16}, 16374 gates)                                                                            
                        time:   [72.490 ms 73.077 ms 73.695 ms]
```
------------------------

Regarding zkVM:
Command: `FILENAME=2023-12-13-15-02-00-op-program-data-log.sh bash run-code.sh`

Before:
```
[2024-01-25T15:14:41Z INFO  kimchi_optimism::mips::witness] processing step=0 pc=00000a99ac insn=000802a667 ips=0.00 pages=4721 mem=18.4 MiB name=_rt0_mips_linux
[2024-01-25T15:14:43Z INFO  kimchi_optimism::mips::witness] Snapshot state in snapshot-state-0.json, step 0
[2024-01-25T15:16:05Z INFO  kimchi_optimism::mips::witness] processing step=10000000 pc=0000382ba8 insn=0000000000 ips=111111.11 pages=4912 mem=19.2 MiB name=gopkg.in/yaml%2ev3.is_blankz
[2024-01-25T15:16:07Z INFO  kimchi_optimism::mips::witness] Snapshot state in snapshot-state-10000000.json, step 10000000
[2024-01-25T15:17:30Z INFO  kimchi_optimism::mips::witness] processing step=20000000 pc=00001e5828 insn=0011000016 ips=114942.53 pages=4993 mem=19.5 MiB name=regexp/syntax.ranges.Swap
[2024-01-25T15:17:32Z INFO  kimchi_optimism::mips::witness] Snapshot state in snapshot-state-20000000.json, step 20000000
[2024-01-25T15:18:55Z INFO  kimchi_optimism::mips::witness] processing step=30000000 pc=000001fa80 insn=0010200018 ips=115830.12 pages=5122 mem=20.0 MiB name=runtime.assertE2I2
[2024-01-25T15:18:57Z INFO  kimchi_optimism::mips::witness] Snapshot state in snapshot-state-30000000.json, step 30000000
```

After:
```
[2024-01-25T07:25:15Z INFO  kimchi_optimism::mips::witness] processing step=10000000 pc=0000382ba8 insn=0000000000 ips=77519.38 pages=4912 mem=19.2 MiB name=gopkg.in/yaml%2ev3.is_blankz                          
[2024-01-25T07:25:17Z INFO  kimchi_optimism::mips::witness] Snapshot state in snapshot-state-10000000.json, step 10000000                                                                                          
[2024-01-25T07:26:46Z INFO  kimchi_optimism::mips::witness] processing step=20000000 pc=00001e5828 insn=0011000016 ips=91324.20 pages=4993 mem=19.5 MiB name=regexp/syntax.ranges.Swap                             
[2024-01-25T07:26:48Z INFO  kimchi_optimism::mips::witness] Snapshot state in snapshot-state-20000000.json, step 20000000                                                                                          
[2024-01-25T07:28:15Z INFO  kimchi_optimism::mips::witness] processing step=30000000 pc=000001fa80 insn=0010200018 ips=97402.60 pages=5122 mem=20.0 MiB name=runtime.assertE2I2                                    
[2024-01-25T07:28:16Z INFO  kimchi_optimism::mips::witness] Snapshot state in snapshot-state-30000000.json, step 30000000                                                                                          
[2024-01-25T07:29:43Z INFO  kimchi_optimism::mips::witness] processing step=40000000 pc=0000185984 insn=00afbf0000 ips=100755.67 pages=5167 mem=20.2 MiB name=encoding/json.unquoteBytes                           
[2024-01-25T07:29:45Z INFO  kimchi_optimism::mips::witness] Snapshot state in snapshot-state-40000000.json, step 40000000                                                                                          
[2024-01-25T07:31:13Z INFO  kimchi_optimism::mips::witness] processing step=50000000 pc=0000119410 insn=008fa10034 ips=102880.66 pages=5196 mem=20.3 MiB name=syscall.read                                         
[2024-01-25T07:31:15Z INFO  kimchi_optimism::mips::witness] Snapshot state in snapshot-state-50000000.json, step 50000000                                                                                          
[2024-01-25T07:32:43Z INFO  kimchi_optimism::mips::witness] processing step=60000000 pc=0000141f0c insn=0024e8ffff ips=104166.67 pages=5227 mem=20.4 MiB name=internal/poll.(*fdMutex).rwunlock                    
[2024-01-25T07:32:45Z INFO  kimchi_optimism::mips::witness] Snapshot state in snapshot-state-60000000.json, step 60000000                                                                                          
[2024-01-25T07:34:12Z INFO  kimchi_optimism::mips::witness] processing step=70000000 pc=000014e9c8 insn=0010200005 ips=105105.11 pages=5260 mem=20.5 MiB name=os.(*File).Read                                      
[2024-01-25T07:34:14Z INFO  kimchi_optimism::mips::witness] Snapshot state in snapshot-state-70000000.json, step 70000000                                                                                          
[2024-01-25T07:35:42Z INFO  kimchi_optimism::mips::witness] processing step=80000000 pc=00001fe84c insn=008c2f00c0 ips=105960.26 pages=5289 mem=20.7 MiB name=golang.org/x/crypto/sha3.keccakF1600                 
[2024-01-25T07:35:43Z INFO  kimchi_optimism::mips::witness] Snapshot state in snapshot-state-80000000.json, step 80000000                                                                                          
[2024-01-25T07:37:12Z INFO  kimchi_optimism::mips::witness] processing step=90000000 pc=0000596ee8 insn=001461fffb ips=106508.88 pages=5316 mem=20.8 MiB name=github.com/ethereum/go-ethereum/trie.newTrieReader   
[2024-01-25T07:37:14Z INFO  kimchi_optimism::mips::witness] Snapshot state in snapshot-state-90000000.json, step 90000000                                                                                          
[2024-01-25T07:38:57Z INFO  kimchi_optimism::mips::witness] processing step=100000000 pc=0000071424 insn=0000000000 ips=105263.16 pages=5349 mem=20.9 MiB name=runtime.reentersyscall                              
[2024-01-25T07:38:59Z INFO  kimchi_optimism::mips::witness] Snapshot state in snapshot-state-100000000.json, step 100000000
```

Which is 10% slower.